Migrate `tests/unit/utils/test_templates.py` to pytest
