+++ tests/pytests/unit/utils/templates/conftest.py
@@ -1,0 +1,9 @@
+import pytest
+
+
+@pytest.fixture
+def render_context():
+    return {
+        "opts": {"cachedir": "/D", "__cli": "salt"},
+        "saltenv": None,
+    }
