+++ tests/pytests/unit/utils/templates/test_jinja.py
@@ -1,41 +1,17 @@
 """
 Tests for salt.utils.templates
 """
-import os
 import re
 
+from collections import OrderedDict
 import pytest
 from salt.exceptions import SaltRenderError
 from salt.utils.templates import render_jinja_tmpl
 
-
-@pytest.fixture
-def minion_opts(tmp_path, minion_opts):
-    minion_opts.update(
-        {
-            "cachedir": str(tmp_path / "jinja-template-cache"),
-            "file_buffer_size": 1048576,
-            "file_client": "local",
-            "file_ignore_regex": None,
-            "file_ignore_glob": None,
-            "file_roots": {"test": [str(tmp_path / "templates")]},
-            "pillar_roots": {"test": [str(tmp_path / "templates")]},
-            "fileserver_backend": ["roots"],
-            "hash_type": "md5",
-            "extension_modules": os.path.join(
-                os.path.dirname(os.path.abspath(__file__)), "extmods"
-            ),
-        }
-    )
-    return minion_opts
+from tests.support.mock import patch
 
 
-@pytest.fixture
-def local_salt():
-    return {}
-
-
-def test_jinja_undefined_error_context(minion_opts, local_salt):
+def test_undefined_error_context(render_context):
     """
     Test that jinja provides both the line number on which the error occurred
     in the Jinja template, and also several lines of context around the error
@@ -63,5 +39,64 @@ def test_jinja_undefined_error_context(minion_opts, local_salt):
     with pytest.raises(SaltRenderError, match=match_regex):
         render_jinja_tmpl(
             jinja_code,
-            dict(opts=minion_opts, saltenv="test", salt=local_salt),
+            render_context,
         )
+
+
+def test_render_sanity(render_context):
+    tmpl = """OK"""
+    res = render_jinja_tmpl(tmpl, render_context)
+    assert res == "OK"
+
+
+def test_render_evaluate(render_context):
+    tmpl = """{{ "OK" }}"""
+    res = render_jinja_tmpl(tmpl, render_context)
+    assert res == "OK"
+
+
+def test_render_evaluate_multi(render_context):
+    tmpl = """{% if 1 -%}OK{%- endif %}"""
+    res = render_jinja_tmpl(tmpl, render_context)
+    assert res == "OK"
+
+
+def test_render_variable(render_context):
+    tmpl = """{{ var }}"""
+    render_context["var"] = "OK"
+    res = render_jinja_tmpl(tmpl, render_context)
+    assert res == "OK"
+
+
+def test_render_tojson_sorted(render_context):
+    templ = """thing: {{ var|tojson(sort_keys=True) }}"""
+    expected = """thing: {"x": "xxx", "y": "yyy", "z": "zzz"}"""
+
+    with patch.dict(render_context, {"var": {"z": "zzz", "y": "yyy", "x": "xxx"}}):
+        res = render_jinja_tmpl(templ, render_context)
+
+    assert res == expected
+
+
+def test_render_tojson_unsorted(render_context):
+    templ = """thing: {{ var|tojson(sort_keys=False) }}"""
+    expected = """thing: {"z": "zzz", "x": "xxx", "y": "yyy"}"""
+
+    # Values must be added to the dict in the expected order. This is
+    # only necessary for older Pythons that don't remember dict order.
+    d = OrderedDict()
+    d["z"] = "zzz"
+    d["x"] = "xxx"
+    d["y"] = "yyy"
+
+    with patch.dict(render_context, {"var": d}):
+        res = render_jinja_tmpl(templ, render_context)
+
+    assert res == expected
+
+
+def test_render_cve_2021_25283(render_context):
+    tmpl = """{{ [].__class__ }}"""
+    render_context["var"] = "OK"
+    with pytest.raises(SaltRenderError):
+        res = render_jinja_tmpl(tmpl, render_context)
